Hello. Just a kind remind to the UI designers.
Very frequently, egroupware UI designers use this kind of table:

<table width=100%>
<tr><td width=1%>Subject</td><td><input name="subject".../></td></tr>

This kind of design causes problems for non-english users. Here <td width=1%> is used to force the first column as narrow as possible, and the second <td> takes up the rest of space. Because English words cannot be broken into two lines, the browser kindly (forced) to enlarge the width of the first column to the length of the word being displayed.

In Chinese, Korean and Japanese, a word can be broken into two lines. Suppose the word "subject" translated to 4 characters/ideographs in Japanese, thus the <td width=1%> would cause this <td> takes up 4 line of space.

I would like to suggest we write the UI like this:
<table style="width:100%">
<!-- It's better to use colgroup here to specify column width.
We don't, thanks to mozilla bug 4510
https://bugzilla.mozilla.org/show_bug.cgi?id=4510 -->
<tr><td style="width: 6em">Subject</td><td><input style="width:100%" name="subject".../></td></tr>

There are several advantages:
1) in most cases 6em is a proper value for the minimum first column width;
2) most Chinese words finish in 6 em space (usually Chinese words are much shorter then English words with same meaning).

This solution is better but not perfect. The perfect solution might be properlly using CSS to render the tabel that automatically expend the second column as wide as possible while still keeping the first column in one line. Also I don't know Japanese, the "most Chinese words finish in 6em space" conclusion may not apply to Japanese text as I saw very long Japanese words before. If there are Japanese people on this list, can you come up and make some better suggestions?